UInt32 to UInt64

+++ src/WorldChat.cpp
@@ -110,7 +110,7 @@ public:
             return false;
 
         Player* player = pChat->GetSession()->GetPlayer();
-        uint32 guid = player->GetGUID();
+        uint64 guid = player->GetGUID();
 
 		if (!player->CanSpeak()){
 			ChatHandler(pChat->GetSession()).PSendSysMessage("[WC] %sYou can't use World Chat while muted!|r", WORLD_CHAT_RED.c_str());
@@ -142,7 +142,7 @@ public:
 
 
             Player* target = itr->second->GetPlayer();
-            uint32 guid2 = target->GetGUID();
+            uint64 guid2 = target->GetGUID();
 
             if (WorldChat[guid2].chat == 1)
             {
@@ -192,7 +192,7 @@ public:
             return false;
 
         Player* player = pChat->GetSession()->GetPlayer();
-        uint32 guid = player->GetGUID();
+        uint64 guid = player->GetGUID();
 
         if (!WorldChat[guid].chat) {
             ChatHandler(player->GetSession()).PSendSysMessage("[WC] %sWorld Chat is disabled. (.chat)|r", WORLD_CHAT_RED.c_str());
@@ -220,7 +220,7 @@ public:
 
 
             Player* target = itr->second->GetPlayer();
-            uint32 guid2 = target->GetGUID();
+            uint64 guid2 = target->GetGUID();
 
             if (WorldChat[guid2].chat == 1 && (target->GetTeamId() == TEAM_HORDE || target->IsGameMaster()))
             {
@@ -244,7 +244,7 @@ public:
             return false;
 
         Player* player = pChat->GetSession()->GetPlayer();
-        uint32 guid = player->GetGUID();
+        uint64 guid = player->GetGUID();
 
         if (!WorldChat[guid].chat) {
             ChatHandler(player->GetSession()).PSendSysMessage("[WC] %sWorld Chat is disabled. (.chat)|r", WORLD_CHAT_RED.c_str());
@@ -272,7 +272,7 @@ public:
 
 
             Player* target = itr->second->GetPlayer();
-            uint32 guid2 = target->GetGUID();
+            uint64 guid2 = target->GetGUID();
 
             if (WorldChat[guid2].chat == 1 && (target->GetTeamId() == TEAM_ALLIANCE || target->IsGameMaster()))
             {
@@ -287,7 +287,7 @@ public:
     static bool HandleWorldChatOnCommand(ChatHandler* handler, const char* /*msg*/)
     {
         Player* player = handler->GetSession()->GetPlayer();
-        uint32 guid = player->GetGUID();
+        uint64 guid = player->GetGUID();
 
         if (!sConfigMgr->GetBoolDefault("World_Chat.Enable", true)) {
             ChatHandler(player->GetSession()).PSendSysMessage("[WC] %sWorld Chat System is disabled.|r", WORLD_CHAT_RED.c_str());
@@ -309,7 +309,7 @@ public:
     static bool HandleWorldChatOffCommand(ChatHandler* handler, const char* /*msg*/)
     {
         Player* player = handler->GetSession()->GetPlayer();
-        uint32 guid = player->GetGUID();
+        uint64 guid = player->GetGUID();
 
         if (!sConfigMgr->GetBoolDefault("World_Chat.Enable", true)) {
             ChatHandler(player->GetSession()).PSendSysMessage("[WC] %sWorld Chat System is disabled.|r", WORLD_CHAT_RED.c_str());
